feat(blog, user): update ID and status fields to use uint types for improved data integrity and add utility functions for string conversion

+++ internal/app/dto/response/blog.go
@@ -3,12 +3,12 @@ package response
 type BlogData struct {
 	Content    string `json:"content"`
 	CreateTime string `json:"create_time"`
-	Id         int64  `json:"id"`
+	Id         uint64 `json:"id"`
 	Problem    struct {
 		ProblemSimpleData
 		ProblemUserScore
 	} `json:"problem,omitempty"`
-	Status     int64          `json:"status"`
+	Status     uint8          `json:"status"`
 	Title      string         `json:"title"`
 	UpdateTime string         `json:"update_time"`
 	User       UserSimpleData `json:"user"`
